+++ packages/dataviews/src/field-types/media.tsx
@@ -8,13 +8,13 @@ export default {
 	render: () => null,
 	Edit: null,
 	sort: () => 0,
-	isValid: {
-		elements: true,
-		custom: () => null,
-	},
 	enableSorting: false,
 	enableGlobalSearch: false,
 	defaultOperators: [],
 	validOperators: [],
 	getFormat: () => ( {} ),
+	// cannot validate any constraint, so
+	// the only available validation for the field author
+	// would be providing a custom validator.
+	validate: {},
 } satisfies FieldType< any >;