+++ packages/ui/src/box/box.tsx
@@ -84,12 +84,15 @@ export const Box = forwardRef< HTMLDivElement, BoxProps >( function Box(
 		backgroundColor,
 		color,
 		padding,
+		borderRadius,
+		borderWidth,
+		borderColor,
 		render = DEFAULT_RENDER,
 		...props
 	},
 	ref
 ) {
-	const style: React.CSSProperties = {};
+	const style: React.CSSProperties = { ...props.style };
 
 	if ( backgroundColor ) {
 		style.backgroundColor = `var(--wpds-color-bg-${ target }-${ backgroundColor }, var(--wpds-color-bg-surface-${ backgroundColor }))`;
@@ -106,5 +109,22 @@ export const Box = forwardRef< HTMLDivElement, BoxProps >( function Box(
 		);
 	}
 
-	return renderElement< 'div' >( render, { style, ...props }, ref );
+	if ( borderRadius ) {
+		style.borderRadius = `var(--wpds-border-radius-${ target }-${ borderRadius }, var(--wpds-border-radius-surface-${ borderRadius }))`;
+	}
+
+	if ( borderWidth ) {
+		style.borderWidth = `var(--wpds-border-width-${ target }-${ borderWidth }, var(--wpds-border-width-surface-${ borderWidth }))`;
+		style.borderStyle = 'solid';
+	}
+
+	if ( borderColor ) {
+		style.borderColor = `var(--wpds-color-stroke-${ target }-${ borderColor }, var(--wpds-color-stroke-surface-${ borderColor }))`;
+	}
+
+	return renderElement< 'div' >( {
+		render,
+		ref,
+		props: { ...props, style },
+	} );
 } );