+++ packages/ui/src/stack/index.ts
@@ -1,0 +1,4 @@
+/**
+ * Internal dependencies
+ */
+export { Stack } from './stack';